refactor(WebAPI): Moved PlexSignIn to vertical slice
