+++ src/PlexApi/PlexApiWrapper.cs
@@ -80,81 +80,6 @@ public class PlexApiWrapper
     }
 
     /// <summary>
-    /// Sign in user with username and password and return user data with Plex authentication token.
-    /// <remarks>NOTE: Plex "Managed" users do not work.</remarks>
-    /// <example>URL: https://plex.tv/api/v2/users/signin?X-Plex-Client-Identifier=Chrome</example>
-    /// </summary>
-    /// <returns></returns>
-    public async Task<Result<PlexAccount>> PlexSignInAsync(PlexAccount plexAccount)
-    {
-        _log.Debug("Requesting PlexToken for account {UserName}", plexAccount.Username);
-
-        var plexTvClient = CreateTvClient();
-
-        var responseResult = await ToResponse(
-            plexTvClient.Authentication.PostUsersSignInDataAsync(
-                new PostUsersSignInDataRequest()
-                {
-                    ClientID = GetClientId,
-                    RequestBody = new PostUsersSignInDataRequestBody()
-                    {
-                        Login = plexAccount.Username,
-                        Password = plexAccount.Password,
-                        RememberMe = false,
-                        VerificationCode = plexAccount.Is2Fa ? plexAccount.VerificationCode : string.Empty,
-                    },
-                }
-            )
-        );
-
-        var result = responseResult.ToApiResult(x => new PlexAccount
-        {
-            Id = plexAccount.Id,
-            DisplayName = plexAccount.DisplayName,
-            Username = plexAccount.Username,
-            Password = plexAccount.Password,
-            IsEnabled = plexAccount.IsEnabled,
-            IsValidated = true,
-            ValidatedAt = DateTime.UtcNow,
-            PlexId = x.UserPlexAccount!.Id,
-            Uuid = x.UserPlexAccount!.Uuid,
-            ClientId = plexAccount.ClientId,
-            Title = x.UserPlexAccount!.Title,
-            Email = x.UserPlexAccount!.Email,
-            HasPassword = x.UserPlexAccount!.HasPassword.GetValueOrDefault(),
-            AuthenticationToken = x.UserPlexAccount!.AuthToken,
-            CustomAuthenticationToken = plexAccount.CustomAuthenticationToken,
-            IsMain = plexAccount.IsMain,
-            PlexAccountServers = [],
-            PlexAccountLibraries = [],
-            Is2Fa = x.UserPlexAccount!.TwoFactorEnabled.GetValueOrDefault(),
-            VerificationCode = string.Empty,
-        });
-
-        if (result.IsSuccess)
-        {
-            _log.Information(
-                "Successfully retrieved the PlexAccount data for user {PlexAccountDisplayName} from the PlexApi",
-                plexAccount.DisplayName
-            );
-        }
-
-        return result;
-    }
-
-    public async Task<Result<string>> RefreshPlexAuthTokenAsync(PlexAccount plexAccount)
-    {
-        var result = await PlexSignInAsync(plexAccount);
-        if (result.IsSuccess)
-        {
-            _log.Debug("Returned token was: {AuthToken}", result.Value.AuthenticationToken);
-            return result.Value.AuthenticationToken;
-        }
-
-        return Result.Fail("Result from RequestPlexSignInDataAsync() was null.").LogError();
-    }
-
-    /// <summary>
     ///     Retrieves all the accessible plex server based on the <see cref="PlexAccount" /> token
     ///     Including the various unique connections to each server.
     ///     <remarks>https://plex.tv/api/v2/resources?X-Plex-Token={{AUTH_TOKEN}}</remarks>
