+++ tests/UnitTests/Application.UnitTests/PlexAccount/GenerateToken/GenerateTokenEndpoint.UnitTests.cs
@@ -17,8 +17,8 @@ public class GenerateTokenEndpointUnitTests : BaseUnitTest
         var seed = await SetupDatabase(2305, config => config.PlexAccountCount = 1);
         var plexAccount = await IDbContext.PlexAccounts.FirstAsync();
 
-        mock.Mock<IPlexApiService>()
-            .Setup(x => x.PlexSignInAsync(It.IsAny<PlexAccount>()))
+        mock.Mock<ICommandDispatch>()
+            .Setup(x => x.ExecuteAsync(It.IsAny<PlexSignInCommand>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync(Result.Ok(FakeData.GetPlexAccount(seed).Generate()));
 
         // Act
@@ -47,8 +47,8 @@ public class GenerateTokenEndpointUnitTests : BaseUnitTest
         await SetupDatabase(232432, config => config.PlexAccountCount = 1);
         var plexAccount = await IDbContext.PlexAccounts.FirstAsync();
 
-        mock.Mock<IPlexApiService>()
-            .Setup(x => x.PlexSignInAsync(It.IsAny<PlexAccount>()))
+        mock.Mock<ICommandDispatch>()
+            .Setup(x => x.ExecuteAsync(It.IsAny<PlexSignInCommand>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync(
                 Result.Fail(new PlexError("Enter verification code") { Code = PlexErrorCodes.EnterVerificationCode })
             );
@@ -78,8 +78,8 @@ public class GenerateTokenEndpointUnitTests : BaseUnitTest
         await SetupDatabase(433222, config => config.PlexAccountCount = 1);
         var plexAccount = await IDbContext.PlexAccounts.FirstAsync();
 
-        mock.Mock<IPlexApiService>()
-            .Setup(x => x.PlexSignInAsync(It.IsAny<PlexAccount>()))
+        mock.Mock<ICommandDispatch>()
+            .Setup(x => x.ExecuteAsync(It.IsAny<PlexSignInCommand>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync(Result.Fail(new PlexError("Unauthorized")).Add401UnauthorizedError());
 
         // Act
@@ -107,8 +107,8 @@ public class GenerateTokenEndpointUnitTests : BaseUnitTest
         await SetupDatabase(433222, config => config.PlexAccountCount = 1);
         var plexAccount = await IDbContext.PlexAccounts.FirstAsync();
 
-        mock.Mock<IPlexApiService>()
-            .Setup(x => x.PlexSignInAsync(It.IsAny<PlexAccount>()))
+        mock.Mock<ICommandDispatch>()
+            .Setup(x => x.ExecuteAsync(It.IsAny<PlexSignInCommand>(), It.IsAny<CancellationToken>()))
             .ReturnsAsync(
                 Result
                     .Fail(new PlexError("Unauthorized"))
